requirements.txt
pandas
ipykernel
matplotlib
# seaborn
tk
pillow
# pandas
plotly
kaleido
seaborn
# matplotlib
